Lift app user state to App.js
In order for app user to be accessible throughout the app, it had to be lifted to the App component and passed down to all screens in the app using the screenProps property of the Navigator component.

State updates are received via a callback which is also added to the screenProps property.